requirements.txt
Python==3.10.19
torch==2.1.0
torchvision==0.16.0
torchaudio==2.1.0
Django==4.2.0
Pillow==10.0.0
numpy==1.24.0